Revert "Move skill files into a side panel"
This reverts commit 26ab55e3f1e5f49121fdfd9cb7235a198b9f0863.
